refactor : 통합/인수 테스트 공통 로직 추가
