+++ server/src/test/java/com/momo/common/acceptance/step/AcceptanceStep.java
@@ -1,4 +1,4 @@
-package com.momo.common.acceptance;
+package com.momo.common.acceptance.step;
 
 import static org.assertj.core.api.Assertions.assertThat;
